[WIP] perf branch: Add valid and spec flags
Add new branch flags that provide speculation information
captured via features such as Last Branch Record (LBR) on
supported hardware.

The valid bit indicates if a recorded taken branch was on
the correct path and the spec bit indicates if the branch
was speculated. Hence, this relies on hardware that also
records taken branches which were speculated but were on
the wrong path and helps capture information about retired
basic blocks. The bits are not set by unsupported hardware.

Suggested-by: Stephane Eranian <eranian@google.com>
Signed-off-by: Sandipan Das <sandipan.das@amd.com>