[INTERNAL] Add boot option to enable amd-pstate driver
The amd-pstate driver is not built by default but we would
like to change that so internal teams can test out the amd-pstate
driver. The catch is that we do not want it to replace the
default acpi-cpufreq driver. To accomplish this a boot command
line option is added to allow users to enable the amd-pstate
driver, the default behavior will be to have the amd-pstate
driver disabled.

To enable the amd-pstate driver add 'amd_pstate.enabled=1' to the
boot command line.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>