[INTERNAL ONLY] perf/x86/amd/lbr: Use freeze bit based on microcode
[FOR INTERNAL USE ONLY - USES NON-PUBLIC INFORMATION]

AMD Last Branch Record Extension Version 2 (LbrExtV2) supports LBR Freeze
on PMI (FLBRI) which is essential for ensuring that branch records remain
consistent with the point of PMU overflow in order to provide a precise
correlation between the two.

The Debug Control register contains the FLBRI bit but its intended position
got swapped with that of the PMC Freeze on PMI (FPMCI) bit due a hardware
bug. While this has been addressed via a microcode fix, not all processor
revisions will receive it and will continue to use the FPMCI bit as FLBRI.
This includes Genoa A0 and A1.

Detect processor revision and microcode level during LBR initialization and
use the correct bit for enabling FLBRI.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>