*debug: x86/fault: Handle RMP faults for kernel addresses
DEBUG-ONLY: this is only here to avoid host crashes in the case of
invalidations for restrictedmem pages not being handled as expected

Well-behaved guests will never switch a page to private state if they
are currently in use for shared host/guest access for things like GHCB
requests, pvclock, etc.

For misbehaved guests, this case needs handling however. Rather than
introducing some sort of mechanism to synchronize the RMP table updates
for a PFN with any in-use kernel mappings to that PFN, simply flip the
PFN back to the shared state that host expects it to be in. This will
unset the validated bit on those pages, so misbehaved guests will see
this in the form of a #VC exception where they can safely terminate.

Currently shared->private page conversions are accompanied with the
removal of the page from the kernel directmap. As such accesses to these
pages will generally result in page-not-found #PF rather than an RMP
fault. There are some race windows where the latter is possible as well,
however, so implement the handler such that it covers both.

Signed-off-by: Michael Roth <michael.roth@amd.com>