*x86/sev: Split direct map for when marking pages private
*squash into (effectively replaces it):
   x86/sev: Invalid pages from direct map when adding it to RMP table

*use the following commit subject/message instead:

When the SEV-SNP RMP table is enabled, host accesses to pages within
the range covered by the RMP table are subject to the following
additional constraints:

  a) for 4K mappings, a write access is only allowed if the PFN is
     marked as shared in the RMP table
  b) for hugepage mappings, a write access is only allowed if all the
     PFNs covered by that mapping are marked as shared in the RMP table

Any violation of these will result in an RMP violation in the form of a
host #PF with the X86_PF_RMP bit set.

The situation described by a) should never happen normally, as there is
never a reason for the host to write to a private page; it would only
result in the guest seeing garbage due to encryption.

The situation described by b) is likely however, since an SEV-SNP guest
will commonly share pages with the host for things like GHCB buffers,
virtio buffers, etc., and since the kernel direct map uses hugepage
mappings by default, kernel accesses that use the direct mapping to
write to these pages would potentially end up overlapping with adjacent
pages that are marked private in the RMP table.

Resolve this issue by splitting the direct map mapping whenever a page
it covers is switched to private in the RMP table.

Currently, this will eventually lead to much of the kernel direct map
being split up over time, since there is no current support for
unsplitting a direct map entry, but that is something that can
potentially be resolved in the future.

Originally-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Michael Roth <michael.roth@amd.com>