*KVM: SEV: fix potential deadlock with dump_ghcb()
Currently dump_ghcb() gets called when sev_es_validate_vmgexit() hits
an error condition, but at that point the GHCB is already mapped, so
dump_ghcb() will cause a deadlock when it attempts to map GHCB page
again.

Pass the GHCB as a pointer argument instead of trying to map it again.

Signed-off-by: Michael Roth <michael.roth@amd.com>