KVM: x86: Add 'fault_is_private' x86 op
This callback is used by the KVM MMU to check whether a #NPF was
or a private GPA or not.

In some cases the full 64-bit error bits for the #NPF will be needed to
make this determination, so also update kvm_mmu_do_page_fault() to
accept the full 64-bit value so it can be plumbed through to the
callback.

Signed-off-by: Michael Roth <michael.roth@amd.com>