*fixup (upm_base_support): KVM: fix restrictedmem GFN range calculations
Invalidations will be issued using inclusive start/end address ranges,
but the current code will calculate exclusive end ranges when clamping
is needed, leading to intended invalidations being issued.

Also, the start/end ranges passed to the notifiers are offsets from the
start of the FD range, 0. Since the memslot's base GFN is bound to the
restrictedmem FD corresponding the the restrictedmem_offset field of
the memslot, this restrictedmem_offset value needs to be subtracted
for start/end in order to get the offset from the base GFN.

Signed-off-by: Michael Roth <michael.roth@amd.com>