[WIP] x86/cpuid: Add smp helper
On some x86 processors, certain functions report different values
depending on the CPU on which the CPUID instruction was executed.

E.g. for AMD Zen 4 processors, the ActiveUmcMask field in leaf
0x80000022 ECX, which provides a way to determine the active memory
controllers, can have different masks on CPUs belonging to different
sockets as each socket can have a different DIMM population scheme.
Each memory channel is assigned a memory controller (UMC) and if no
DIMMs are attached to a channel, the corresponding memory controller
is inactive.

Add a new helper that executes CPUID on a particular CPU and returns
the EAX, EBX, ECX and EDX register values.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>