[WIP] perf/x86/amd/uncore: Add memory controller event support
Unified Memory Controller (UMC) events were introduced with Zen 4 as a
part of the Performance Monitoring Version 2 (PerfMonV2) enhancements.
Each socket supports upto 12 channels of DDR5, each controlled by a set
of dedicated UMCs. Since the PERF_CTL and PERF_CTR MSRs share the same
address space within each socket, sockets are considered to be channel
groups. This can change for future products and in that case, the socket
identifier cannot be used to distinguish between channels groups.

Because of channel groups, these uncore units are usable exclusively
from a specific set of online CPUs, in this case, the ones belonging to
the same socket. For a given channel group, CPUs outside the group do
not need to maintain an uncore context. Hence, an additional "owner"
attribute is required for managing exclusivity.

Each UMC has its own set of counters. The number of active UMCs (depends
on the DIMM population scheme) and the number of counters for each UMC
can be determined from CPUID 0x80000022 EBX and ECX.

Signed-off-by: Sandipan Das <sandipan.das@amd.com>