[PATCH] [INTERNAL] drivers/misc: Page fault hooks for hardware tracing.
To assist with hardware debugging so that their buffers are not filled
during the handling of a page fault, this patch provides hooks on the
entry and exit of the page fault handler to inform hardware that we
are entering/exiting the page fault handler. This will allow hardware
to selectively disable their logging mechanisms while the page fault
handler runs.

To control this behavior a sysfs file is created,
/sys/kernel/mm/hw_pf_tracing, to provide controls to enable the
notifications and select which SMT thread to notify on.

Additionally, there is a sysfs file (/sys/kernel/mm/hw_pf_stats)
that will display per-cpu page fault hits while this is enabled.

Signed-off-by: Nathan Fontenot <nathan.fontenot@amd.com>