KVM: x86: Add platform hook for initializing private memory
TODO: add a CONFIG option rather than relying on a __weak reference for
arch/platform that don't implement this hook

All gmem pages are expected to be 'private' as defined by a particular
arch/platform. Platforms like SEV-SNP require additional operations to
move these pages into a private state, so implement a hook that can be
used to prepare this memory prior to mapping it into a guest.

In some cases, such as SEV-SNP, this memory needs to be written to prior
to converting it to a private page, so also add a __kvm_gmem_get_pfn()
helper that can be used to skip the hook.

Signed-off-by: Michael Roth <michael.roth@amd.com>