KVM: x86: Add platform hooks for private memory invalidations
TODO: add a CONFIG option that can be to completely skip arch
invalidation loop and avoid __weak references for arch/platforms that
don't need an additional invalidation hook.

In some cases, like with SEV-SNP, guest memory needs to be updated in a
platform-specific manner before it can be safely freed back to the host.
Add hooks to wire up handling of this sort to the invalidation notifiers
for gmem.

Also issue invalidations of all allocated pages during memslot unbinding
so that the pages are not left in an unusable state when they eventually
get freed back to the host upon FD release.

Signed-off-by: Michael Roth <michael.roth@amd.com>