x86/resctrl: Move default group file creation during mount
Currently, the resctrl default group files are created during kernel
init and rest of the files are added during mount. If the new files
needed to be added to the default group during mount then it has to
be done separately again.

This can avoided if all the files are created during the mount.
So, move default group file creation while mounting.

Signed-off-by: Babu Moger <babu.moger@amd.com>