*fixup: Handle non-zero cur_page field when entry is part of contiguous range
Current code handle non-zero 'cur_page' fields when they are at the
beginning of a new GPA range, but it is not handling properly when
iterating through subsequent entries which are part of a contiguous
range. Fix up the handling so they entries are not combined into a
larger contiguous range and are instead processed later as the start
of a new contiguous range.

Signed-off-by: Michael Roth <michael.roth@amd.com>